requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# OpenAI and ML dependencies
openai==1.3.7
pandas==2.1.4
numpy==1.24.3

# Database
# sqlite3 is built into Python, no need to install

# Dashboard and visualization
streamlit==1.28.1
plotly==5.17.0

# File handling
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For production deployment
gunicorn==21.2.0 